pyproject.toml
[tool.poetry]
name = "wagtail_tenants"
version = "0.1.9"
description = "Adds multitenancy based on django_tenants to wagtail cms"
authors = ["Boris Brue <boris@zuckersalzundpfeffer.de>"]
license = "MIT"
readme = "README.md"
homepage = "https://wagtail-tenants.readthedocs.io/en/latest/"
repository = "https://github.com/borisbrue/wagtail-tenants"

[tool.poetry.dependencies]
python = "^3.7.0"
django-tenants = "^3.3.4"
django-dbbackup = "^3.3.0"


[tool.poetry.dev-dependencies]
black = "22.3.0"
flake8 = "^4.0.1"
sphinx = {version = "^3", optional = true}
wagtail = "^4.0.2"
isort = "^5.10.1"
flake8-black = "0.3.3"
pytest-django = "^4.5.2"
tox = "^3.24.4"
pytest-xdist = "^2.4.0"
factory-boy = "^3.2.1"
pytest = "^6.2.5"
coverage = "^6.2"
myst-parser = "^0.15.2"
sphinx-rtd-theme = "^1.0.0"
psycopg2 = "^2.9.2"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
docs = ["sphinx"]